+++ src/compiler/semantic.rs
@@ -136,9 +136,12 @@ mod tests {
             statements: vec![Statement::Impl {
                 trait_name: None,
                 type_name: "Point".to_string(),
+            type_params: Vec::new(),
+            type_param_bounds: Vec::new(),
                 methods: vec![Statement::Function {
                     name: "draw".to_string(),
             type_params: Vec::new(),
+            type_param_bounds: Vec::new(),
                     params: vec![(
                         "self".to_string(),
                         DataType::StructNamed("Point".to_string()),